--- delphi/GrFN/ProgramAnalysisGraph.py
@@ -1,6 +1,0 @@
-import networkx as nx
-
-
-class ProgramAnalysisGraph(nx.DiGraph):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
