--- delphi/analysis/sensitivity/emulators.py
@@ -1,10 +1,0 @@
-from abc import ABCMeta
-
-
-class EmulationAnalyzer(metaclass=ABCMeta):
-    """
-    Meta-class for all emulation based sensitivity analysis methods
-    """
-
-    def __init__(self):
-        pass
