--- delphi/analysis/sensitivity/variogram_methods.py
@@ -1,10 +1,0 @@
-from abc import ABCMeta
-
-
-class VariogramAnalyzer(metaclass=ABCMeta):
-    """
-    Meta-class for all Variogram methods of sensitivity analysis
-    """
-
-    def __init__(self):
-        pass
