--- scripts/program_analysis/PETPT_numpy.json
@@ -1,357 +1,0 @@
-{
-  "functions": [
-    {
-      "name": "petpt__assign__td_1",
-      "type": "assign",
-      "reference": 14,
-      "sources": [
-        {
-          "name": "tmax",
-          "type": "variable"
-        },
-        {
-          "name": "tmin",
-          "type": "variable"
-        }
-      ],
-      "target": "td"
-    },
-    {
-      "name": "petpt__decision__albedo_1",
-      "type": "decision",
-      "target": "albedo",
-      "reference": 15,
-      "sources": [
-        {
-          "name": "xhlai",
-          "type": "variable"
-        },
-        {
-          "name": "msalb",
-          "type": "variable"
-        }
-      ]
-    },
-    {
-      "name": "petpt__assign__slang_1",
-      "type": "assign",
-      "reference": 19,
-      "sources": [
-        {
-          "name": "srad",
-          "type": "variable"
-        }
-      ],
-      "target": "slang"
-    },
-    {
-      "name": "petpt__assign__eeq_1",
-      "type": "assign",
-      "reference": 20,
-      "sources": [
-        {
-          "name": "slang",
-          "type": "variable"
-        },
-        {
-          "name": "albedo",
-          "type": "variable"
-        },
-        {
-          "name": "td",
-          "type": "variable"
-        }
-      ],
-      "target": "eeq"
-    },
-    {
-      "name": "petpt__assign__eo_0",
-      "type": "assign",
-      "reference": 21,
-      "sources": [
-        {
-          "name": "eeq",
-          "type": "variable"
-        }
-      ],
-      "target": "eo"
-    },
-    {
-      "name": "petpt__decision__eo_1",
-      "type": "decision",
-      "target": "eo",
-      "reference": 25,
-      "sources": [
-        {
-          "name": "tmax",
-          "type": "variable"
-        },
-        {
-          "name": "eeq",
-          "type": "variable"
-        },
-        {
-          "name": "eo_0",
-          "type": "variable"
-        }
-      ]
-    },
-    {
-      "name": "petpt__decision__eo_2",
-      "type": "decision",
-      "target": "eo",
-      "reference": 22,
-      "sources": [
-        {
-          "name": "tmax",
-          "type": "variable"
-        },
-        {
-          "name": "eeq",
-          "type": "variable"
-        },
-        {
-          "name": "eo_1",
-          "type": "variable"
-        }
-      ]
-    },
-    {
-      "name": "petpt__assign__eo_3",
-      "type": "assign",
-      "reference": 27,
-      "sources": [
-        {
-          "name": "eo",
-          "type": "variable"
-        }
-      ],
-      "target": "eo"
-    },
-    {
-      "name": "petpt",
-      "type": "container",
-      "input": [
-        {
-          "name": "msalb",
-          "domain": "real"
-        },
-        {
-          "name": "srad",
-          "domain": "real"
-        },
-        {
-          "name": "tmax",
-          "domain": "real"
-        },
-        {
-          "name": "tmin",
-          "domain": "real"
-        },
-        {
-          "name": "xhlai",
-          "domain": "real"
-        },
-        {
-          "name": "eo",
-          "domain": "real"
-        }
-      ],
-      "variables": [
-        {
-          "name": "msalb",
-          "domain": "real"
-        },
-        {
-          "name": "srad",
-          "domain": "real"
-        },
-        {
-          "name": "tmax",
-          "domain": "real"
-        },
-        {
-          "name": "tmin",
-          "domain": "real"
-        },
-        {
-          "name": "xhlai",
-          "domain": "real"
-        },
-        {
-          "name": "eo",
-          "domain": "real"
-        },
-        {
-          "name": "albedo",
-          "domain": "real"
-        },
-        {
-          "name": "eeq",
-          "domain": "real"
-        },
-        {
-          "name": "slang",
-          "domain": "real"
-        },
-        {
-          "name": "td",
-          "domain": "real"
-        },
-        {
-          "name": "IF_1",
-          "domain": "boolean"
-        },
-        {
-          "name": "IF_2",
-          "domain": "boolean"
-        },
-        {
-          "name": "IF_3",
-          "domain": "boolean"
-        }
-      ],
-      "body": [
-        {
-          "name": "petpt__assign__td_1",
-          "output": {
-            "variable": "td",
-            "index": 1
-          },
-          "input": [
-            {
-              "variable": "tmax",
-              "index": 0
-            },
-            {
-              "variable": "tmin",
-              "index": 0
-            }
-          ]
-        },
-        {
-          "name": "petpt__decision__albedo_1",
-          "output": {
-            "variable": "albedo",
-            "index": 1
-          },
-          "input": [
-            {
-              "variable": "xhlai",
-              "index": 0
-            },
-            {
-              "variable": "msalb",
-              "index": 0
-            }
-          ]
-        },
-        {
-          "name": "petpt__assign__slang_1",
-          "output": {
-            "variable": "slang",
-            "index": 1
-          },
-          "input": [
-            {
-              "variable": "srad",
-              "index": 0
-            }
-          ]
-        },
-        {
-          "name": "petpt__assign__eeq_1",
-          "output": {
-            "variable": "eeq",
-            "index": 1
-          },
-          "input": [
-            {
-              "variable": "slang",
-              "index": 1
-            },
-            {
-              "variable": "albedo",
-              "index": 1
-            },
-            {
-              "variable": "td",
-              "index": 1
-            }
-          ]
-        },
-        {
-          "name": "petpt__assign__eo_0",
-          "output": {
-            "variable": "eo",
-            "index": 0
-          },
-          "input": [
-            {
-              "variable": "eeq",
-              "index": 1
-            }
-          ]
-        },
-        {
-          "name": "petpt__decision__eo_1",
-          "output": {
-            "variable": "eo",
-            "index": 1
-          },
-          "input": [
-            {
-              "variable": "tmax",
-              "index": 0
-            },
-            {
-              "variable": "eeq",
-              "index": 1
-            },
-            {
-              "variable": "eo",
-              "index": 0
-            }
-          ]
-        },
-        {
-          "name": "petpt__decision__eo_2",
-          "output": {
-            "variable": "eo",
-            "index": 2
-          },
-          "input": [
-            {
-              "variable": "tmax",
-              "index": 0
-            },
-            {
-              "variable": "eeq",
-              "index": 1
-            },
-            {
-              "variable": "eo",
-              "index": 1
-            }
-          ]
-        },
-        {
-          "name": "petpt__assign__eo_3",
-          "output": {
-            "variable": "eo",
-            "index": 3
-          },
-          "input": [
-            {
-              "variable": "eo",
-              "index": 2
-            }
-          ]
-        }
-      ]
-    }
-  ],
-  "start": "petpt",
-  "name": "PETPT.json",
-  "dateCreated": "2019-03-09"
-}
